spectr_utmc/controller_snapshot/home/voicelink/.releases/1.4.12/web_admin/requirements.txt
Babel==2.8.0
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Limiter==1.2.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
IPy==1.00
Jinja2==2.11.1
MarkupSafe==1.1.1
SQLAlchemy==1.3.13
Tempita==0.5.2
WTForms==2.2.1
Werkzeug==1.0.0
blinker==1.4
click==7.1.1
decorator==4.4.2
itsdangerous==1.1.0
limits==1.5.1
pbr==5.4.5
pytz==2019.3
six==1.14.0
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
typing==3.7.4.1
semantic-version=2.8.5
monotonic=1.6
